Some mpi implementations and Gate X parallelized with threads

+++ src/executor.hpp
@@ -19,7 +19,7 @@ public:
     void apply_gate(const std::string& gate_name, const std::vector<int>& qubits);
     void apply_parametric_gate(const std::string& gate_name, const std::vector<int>& qubits, std::vector<double>& param);
     void apply_unitary(const std::string& gate_name, const Matrix& matrix, const std::vector<int>& qubits);
-    int get_nonzero_position();
+    uint64_t get_nonzero_position();
     void restart_statevector();
     ResultCunqa run(QuantumCircuit& quantumcircuit, int shots = 10);
